feat: move util function out of component
