+++ components/Transaction/AddressTransactions.tsx
@@ -7,7 +7,8 @@ import BCHIcon from 'assets/bch-logo.png'
 import EyeIcon from 'assets/eye-icon.png'
 import CheckIcon from 'assets/check-icon.png'
 import XIcon from 'assets/x-icon.png'
-import TableContainer, { compareNumericString } from '../../components/TableContainer'
+import TableContainer from '../../components/TableContainer'
+import { compareNumericString } from 'utils/index'
 import moment from 'moment'
 
 interface IProps {
