+++ pages/payments/index.tsx
@@ -3,14 +3,14 @@ import supertokensNode from 'supertokens-node'
 import * as SuperTokensConfig from '../../config/backendConfig'
 import Session from 'supertokens-node/recipe/session'
 import { GetServerSideProps } from 'next'
-import TableContainer, { compareNumericString } from '../../components/TableContainer'
+import TableContainer from '../../components/TableContainer'
 import { ButtonDisplayData } from 'redis/dashboardCache'
 import Image from 'next/image'
 import Link from 'next/link'
 import XECIcon from 'assets/xec-logo.png'
 import BCHIcon from 'assets/bch-logo.png'
 import EyeIcon from 'assets/eye-icon.png'
-import { formatQuoteValue } from 'utils/index'
+import { formatQuoteValue, compareNumericString } from 'utils/index'
 import { XEC_NETWORK_ID, USD_QUOTE_ID } from 'constants/index'
 import moment from 'moment'
